fix(sqlserver): align ReleaseLockAsync sentinel and timeline count type with PG
- ReleaseLockAsync: use 1900-01-01 UTC sentinel matching InitializeAsync
  (was DateTime.MinValue with Unspecified Kind)
- _GetTimelineStats: read COUNT with GetInt64 then cast to int, matching
  PostgreSQL's safer pattern

+++ src/Headless.Messaging.SqlServer/SqlServerMonitoringApi.cs
@@ -290,7 +290,7 @@ internal sealed class SqlServerMonitoringApi(
 
                         while (await reader.ReadAsync(ct).ConfigureAwait(false))
                         {
-                            dictionary.Add(reader.GetString(0), reader.GetInt32(1));
+                            dictionary.Add(reader.GetString(0), (int)reader.GetInt64(1));
                         }
 
                         return dictionary;
